pyproject.toml
[project]
name = "skeletonservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.3",
    "uvicorn>=0.41.0",
]
